♻️ [refactor] : constants 파일에 있던 interface 분리
